+++ demo/components/ComponentsScreen/ComponentsScreen.js
@@ -2,8 +2,8 @@
 import React, { createElement } from 'react'
 import { Route } from 'react-router-dom'
 
-import * as libProps from 'utils/lib-props'
-import routesMap, { componentRoutes } from 'utils/routes-map'
+import * as libProps from 'lib/lib-props'
+import routesMap, { componentRoutes } from 'lib/routes-map'
 
 import GroupNav from 'components/universal/GroupNav'
 import Header from 'components/universal/Header'
@@ -16,6 +16,24 @@ type Props = {
 
 const { subRoutes } = routesMap.components
 
+// Utility transforms an object of props into a string representation
+const renderPropsText = props => {
+  let text = ''
+
+  Object.keys(props).forEach(prop => {
+    const value = props[prop]
+    switch (typeof value) {
+      case 'boolean':
+        if (value) text += ` ${prop}`
+        break
+      default:
+        if (value !== null) text += ` ${prop}="${value}"`
+    }
+  })
+
+  return text
+}
+
 export default ({ location: { state } }: Props) => (
   <div className="grid-container columns-page-layout m-5">
     <div className="guides">
@@ -25,7 +43,12 @@ export default ({ location: { state } }: Props) => (
         <Route
           key={route}
           path={subRoutes[route].pathname}
-          render={() => createElement(routesMap.components.screens[route], libProps)}
+          render={() =>
+            createElement(routesMap.components.screens[route], {
+              ...libProps,
+              renderPropsText,
+            })
+          }
         />
       ))}
     </div>