+++ src/Button/Button.js
@@ -5,14 +5,28 @@ import type { ThemeColors } from '../types'
 export type Props = {
   /** Creates a full-width button */
   block?: boolean,
-  /** Theme color used to compute BS color class */
-  color?: ThemeColors | 'link',
-  /** Computes the button as anchor class */
+  /** Sets button background or link font color */
+  color?: ThemeColors | 'link' | 'body',
+  /** Sets button to display as a link */
   link?: boolean,
   /** Creates outline style button, uses `color` for outline theme. */
   outline?: boolean,
   /** Create a small or large style button */
   size?: 'small' | 'large',
+  // --- Text props
+  /** Sets font size inline style */
+  fontSize?: number | string,
+  /** Sets font weight class */
+  fontWeight: 'light' | 'normal' | 'bold',
+  /** Sets letter spacing inline style */
+  letterSpacing?: number | string,
+  /** Sets line height inline style */
+  lineHeight?: number | string,
+  /** Sets text alignment class */
+  textAlign?: 'justify' | 'right' | 'center' | 'left',
+  italic?: boolean,
+  monospace?: boolean,
+  uppercase?: boolean,
 }
 
 const Button = ({
@@ -20,12 +34,15 @@ const Button = ({
   color,
   fontSize,
   fontWeight,
+  italic = false,
   letterSpacing,
   lineHeight,
   link = false,
+  monospace = false,
   outline = false,
   size,
   textAlign,
+  uppercase = false,
   ...rest
 }: Props) =>
   componentryElem({
@@ -41,8 +58,15 @@ const Button = ({
       'btn-sm': size === 'small',
       'btn-lg': size === 'large',
       disabled: rest.disabled,
+      // --- Typography
+      'font-italic': italic,
+      'text-monospace': monospace,
+      'text-uppercase': uppercase,
       [`font-weight-${fontWeight}`]: fontWeight,
       [`text-${textAlign}`]: textAlign,
+      // When rendering a link style button, a passed color should style the
+      // color of the link
+      [`text-${color}`]: link && color,
     },
     ...rest,
   })